Codebase/airflow/requirements.txt
apache-airflow==2.7.0
apache-airflow-providers-amazon
apache-airflow-providers-snowflake
beautifulsoup4==4.12.2
boto3
chromedriver-py==121.0.6167.85
cryptography==41.0.3
dbt-snowflake==1.5.0
langchain==0.0.350
langchain-openai==0.0.5
lxml==4.9.3
pandas==1.5.3
pinecone-client
psycopg2-binary==2.9.7
pymupdf
python-dotenv==1.0.0
requests==2.31.0
selenium==4.15.2
snowflake-connector-python
urllib3==1.26.18
webdriver_manager
